# swagger.json
{"openapi":"3.0.0","paths":{"/session/login":{"post":{"operationId":"SessionController_postLogin","summary":"Post the login credentials data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the jwt data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["session"]}},"/user/get":{"get":{"operationId":"UserController_getUser","summary":"Get the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/get/professionals":{"get":{"operationId":"UserController_getUsersByCrp","summary":"Get all the data from the professionals","parameters":[],"responses":{"200":{"description":"Returns a JSON with the data from all the professionals","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsersByCrpResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"]}},"/user/get/patients":{"get":{"operationId":"UserController_getUsersByPatientId","summary":"Get all the data from the patients","parameters":[],"responses":{"200":{"description":"Returns a JSON with the data from all the patients","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsersByPatientIdResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/post":{"post":{"operationId":"UserController_postUser","summary":"Post the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostUserReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"]}},"/user/put":{"put":{"operationId":"UserController_putUser","summary":"Put the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutUserReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/delete":{"delete":{"operationId":"UserController_deleteUser","summary":"Delete the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}}},"info":{"title":"ingt-ms-user","description":"Rational Emotive Behaviour Therapy - Dysfunctional Thought Record","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/ingt-ms-user"}],"components":{"securitySchemes":{"auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","in":"header"}},"schemas":{"PostLoginReqDto":{"type":"object","properties":{"email":{"type":"string"},"crp":{"type":"string"},"password":{"type":"string"}},"required":["email","crp","password"]},"PostLoginResDto":{"type":"object","properties":{"jwt":{"type":"string"}},"required":["jwt"]},"ErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"birthdate":{"type":"string"},"crp":{"type":"string"},"responsibleCrp":{"type":"string"},"patientId":{"type":"number"}},"required":["name","address","phone","email","birthdate","crp","responsibleCrp","patientId"]},"GetUserResDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"}},"required":["user"]},"GetUsersByCrpResDto":{"type":"object","properties":{"user":{"type":"array","items":{"type":"string"}}},"required":["user"]},"GetUsersByPatientIdResDto":{"type":"object","properties":{"user":{"type":"array","items":{"type":"string"}}},"required":["user"]},"PostUserReqDto":{"type":"object","properties":{"cpfCnpj":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"birthdate":{"type":"string"},"crp":{"type":"string"},"responsibleCrp":{"type":"string"},"password":{"type":"string"}},"required":["cpfCnpj","name","address","phone","email","birthdate","crp","responsibleCrp","password"]},"PutUserReqDto":{"type":"object","properties":{"address":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"}},"required":["address","phone","password"]},"DeleteUserResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}